sched/fair: do not reset target_capacity
Currently, we reset target_capacity to ULONG_MAX in placement boost.
This causes us to bypass that capacity check in the cpu loop thereby
allowing us to visit higher capacity cluster.

Instead choose to keep target_capacity intact and delete the capacity
check in cpu loop.

The capacity check was introduced by
commit cf9d0b63a3dd6ae051412b401c9d7499deeb43e0 ("sched/fair: fix
incorrect CPU selection for non latency sensitive tasks")
and it is safe to delete it because once we find a cpu in a group we
exit the group traversing loop. We will never have an situation
when we have found a cpu and are visiting higher capacity cluster,
except for placement boost when we intentionally want to visit higher
capacity cpus and were using the target_capacity = ULONG_MAX trick.

Change-Id: I73d25a1e7b30c0420e175371bd9f5911c71ddee0
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
[Fixup minor merge errors]
Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: MadeOfGreat <ravenklawasd@gmail.com>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>
Signed-off-by: DarkAngelGR <asavvo01@gmail.com>